PIMS-2627 Update project when changing milestone status
