+++ backend/dal/Services/Concrete/ProjectService.cs
@@ -398,195 +398,7 @@ namespace Pims.Dal.Services
                 if (originalProject.Tasks.Any(t => !t.Task.IsOptional && !t.IsCompleted && incompleteStatusTaskIds.Contains(t.TaskId))) throw new InvalidOperationException("Not all required tasks have been completed.");
             }
 
-            // Update a project
-            this.Context.Entry(originalProject).CurrentValues.SetValues(project);
-            this.Context.SetOriginalRowVersion(originalProject);
-
-            foreach (var property in project.Properties)
-            {
-                var existingProperty = originalProject.Properties.FirstOrDefault(b => b.PropertyType == PropertyTypes.Land
-                && b.ParcelId == property.ParcelId
-                && b.ProjectId == project.Id
-                ||
-                b.PropertyType == PropertyTypes.Building
-                && b.ProjectId == project.Id
-                && b.BuildingId == property.BuildingId);
-
-                if (existingProperty == null)
-                {
-                    //Todo: Navigation properties on project object were causing concurrency exceptions.
-                    var eproperty = property.PropertyType == PropertyTypes.Land ? this.Context.Parcels.Find(property.ParcelId) : this.Context.Buildings.Find(property.BuildingId) as Property;
-                    // Ignore properties that don't exist.
-                    if (eproperty != null)
-                    {
-                        if (property.PropertyType == PropertyTypes.Land)
-                        {
-                            var existingParcel = this.Context.Parcels
-                                .Include(p => p.Agency)
-                                .Include(p => p.Evaluations)
-                                .Include(p => p.Fiscals)
-                                .FirstOrDefault(p => p.Id == property.ParcelId);
-                            originalProject.ThrowIfPropertyNotInProjectAgency(existingParcel);
-                            if (existingParcel.ProjectNumber != null) throw new InvalidOperationException("Parcels in a Project cannot be added to another Project.");
-                            existingParcel.ProjectNumber = project.ProjectNumber;
-                        }
-                        else
-                        {
-                            var existingBuilding = this.Context.Buildings
-                                .Include(b => b.Agency)
-                                .Include(p => p.Evaluations)
-                                .Include(p => p.Fiscals)
-                                .FirstOrDefault(p => p.Id == property.BuildingId);
-                            originalProject.ThrowIfPropertyNotInProjectAgency(existingBuilding);
-                            if (existingBuilding.ProjectNumber != null) throw new InvalidOperationException("Buildings in a Project cannot be added to another Project.");
-                            existingBuilding.ProjectNumber = project.ProjectNumber;
-                        }
-                        originalProject.AddProperty(eproperty);
-                    }
-                }
-                else
-                {
-                    if (property.PropertyType == PropertyTypes.Land)
-                    {
-                        // Only allow editing the classification and evaluations/fiscals for now
-                        existingProperty.Parcel.ProjectNumber = originalProjectNumber;
-                        existingProperty.Parcel.ClassificationId = property.Parcel.ClassificationId;
-                        existingProperty.Parcel.Zoning = property.Parcel.Zoning;
-                        existingProperty.Parcel.ZoningPotential = property.Parcel.ZoningPotential;
-                        foreach (var evaluation in property.Parcel.Evaluations)
-                        {
-                            var existingEvaluation = existingProperty.Parcel.Evaluations
-                                .FirstOrDefault(e => e.Date == evaluation.Date && e.Key == evaluation.Key);
-
-                            if (existingEvaluation == null)
-                            {
-                                existingProperty.Parcel.Evaluations.Add(evaluation);
-                            }
-                            else
-                            {
-                                this.Context.Entry(existingEvaluation).CurrentValues.SetValues(evaluation);
-                            }
-                        }
-                        foreach (var fiscal in property.Parcel.Fiscals)
-                        {
-                            var existingFiscal = existingProperty.Parcel.Fiscals
-                                .FirstOrDefault(e => e.FiscalYear == fiscal.FiscalYear && e.Key == fiscal.Key);
-
-                            if (existingFiscal == null)
-                            {
-                                existingProperty.Parcel.Fiscals.Add(fiscal);
-                            }
-                            else
-                            {
-                                this.Context.Entry(existingFiscal).CurrentValues.SetValues(fiscal);
-                            }
-                        }
-                    }
-                    else if (property.PropertyType == PropertyTypes.Building)
-                    {
-                        // Only allow editing the classification and evaluations/fiscals for now
-                        existingProperty.Building.ProjectNumber = originalProjectNumber;
-                        existingProperty.Building.ClassificationId = property.Building.ClassificationId;
-                        foreach (var evaluation in property.Building.Evaluations)
-                        {
-                            var existingEvaluation = existingProperty.Building.Evaluations
-                                .FirstOrDefault(e => e.Date == evaluation.Date && e.Key == evaluation.Key);
-
-                            if (existingEvaluation == null)
-                            {
-                                existingProperty.Building.Evaluations.Add(evaluation);
-                            }
-                            else
-                            {
-                                this.Context.Entry(existingEvaluation).CurrentValues.SetValues(evaluation);
-                            }
-                        }
-                        foreach (var fiscal in property.Building.Fiscals)
-                        {
-                            var existingFiscal = existingProperty.Building.Fiscals
-                                .FirstOrDefault(e => e.FiscalYear == fiscal.FiscalYear && e.Key == fiscal.Key);
-
-                            if (existingFiscal == null)
-                            {
-                                existingProperty.Building.Fiscals.Add(fiscal);
-                            }
-                            else
-                            {
-                                this.Context.Entry(existingFiscal).CurrentValues.SetValues(fiscal);
-                            }
-                        }
-                    }
-                }
-            }
-
-            // Remove any properties from this project that are no longer associated.
-            var removePropertyIds = originalProject.Properties.Select(p => p.Id).Except(project.Properties.Select(p => p.Id));
-            var removeProperties = originalProject.Properties.Where(p => removePropertyIds.Contains(p.Id));
-
-            var removeParcelIds = removeProperties.Where(p => p.ParcelId.HasValue).Select(p => p.ParcelId.Value).ToArray();
-            var removeParcels = this.Context.Parcels.Where(p => removeParcelIds.Contains(p.Id));
-            removeParcels.ForEach(p =>
-            {
-                p.ProjectNumber = null;
-                this.Context.Parcels.Update(p);
-            });
-
-            var removeBuildingIds = removeProperties.Where(b => b.BuildingId.HasValue).Select(p => p.BuildingId.Value).ToArray();
-            var removeBuildings = this.Context.Buildings.Where(p => removeBuildingIds.Contains(p.Id));
-            removeBuildings.ForEach(b =>
-            {
-                b.ProjectNumber = null;
-                this.Context.Buildings.Update(b);
-            });
-
-            originalProject.Properties.RemoveAll(p => removePropertyIds.Contains(p.Id));
-
-            foreach (var task in project.Tasks)
-            {
-                var originalProjectTask = originalProject.Tasks.FirstOrDefault(t => t.TaskId == task.TaskId);
-
-                if (originalProjectTask == null)
-                {
-                    originalProject.Tasks.Add(task);
-                }
-                else
-                {
-                    this.Context.Entry(originalProjectTask).CurrentValues.SetValues(task);
-                }
-            }
-
-            foreach (var response in project.Responses)
-            {
-                var originalProjectResponse = originalProject.Responses.FirstOrDefault(r => r.AgencyId == response.AgencyId);
-
-                if (originalProjectResponse == null)
-                {
-                    originalProject.Responses.Add(response);
-                }
-                else
-                {
-                    this.Context.Entry(originalProjectResponse).CurrentValues.SetValues(response);
-                }
-            }
-
-            var status = this.Context.ProjectStatus
-                .Include(s => s.Tasks)
-                .FirstOrDefault(s => s.Id == project.StatusId);
-            project.Status = status;
-
-            // If the tasks haven't been specified, generate them.
-            var taskIds = project.Tasks.Select(t => t.TaskId).ToArray();
-            // Add the tasks for project status if they are not already added.
-            foreach (var task in status.Tasks.Where(t => !taskIds.Contains(t.Id)))
-            {
-                originalProject.Tasks.Add(new ProjectTask(project, task));
-            }
-
-            // Update project financials if the project is still active.
-            if (!project.IsProjectClosed(_options.Project))
-            {
-                originalProject.UpdateProjectFinancials();
-            }
+            originalProject.Merge(project, this.Context, _options.Project);
 
             this.Context.SaveChanges();
             this.Context.CommitTransaction();
@@ -728,6 +540,8 @@ namespace Pims.Dal.Services
             var statusTaskIds = fromStatus.Tasks.Where(t => !t.IsOptional).Select(t => t.Id);
             var incompleteStatusTaskIds = statusTaskIds.Any() && (project.Tasks.Any() || originalProject.Tasks.Any()) ? incompleteTaskIds.Intersect(statusTaskIds) : statusTaskIds;
             if (incompleteStatusTaskIds.Any()) throw new InvalidOperationException("Not all required tasks have been completed.");
+
+            originalProject.Merge(project, this.Context, _options.Project);
 
             // Hardcoded logic to handle milestone project status transitions.
             // This could be extracted at some point to a configurable layer, but not required presently.
